Added .pkgmeta file and updated BetterBags definition
A new .pkgmeta file has been added to the project with specific package configurations. The README.md and *.Definition.lua files are set to be ignored. Also, the 'BetterBags' definition file has been renamed for clarity and a new function 'AddPluginConfig' has been introduced to add plugin's configuration to the BetterBags configuration.