Introduced CI/CD workflow for Retail project
A new continuous integration and deployment (CI/CD) workflow has been added to the Retail project. This workflow triggers on push events to any tag or the main branch. It runs on an Ubuntu-latest environment and uses several secrets for API keys and OAuth tokens. The steps in this workflow include cloning the project and packaging & releasing it using a specific GitHub Action from BigWigsMods/packager.